% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach-duckdb.R
\name{attach_duckdb}
\alias{attach_duckdb}
\title{Attach an Additional DuckDB File to Connection}
\usage{
attach_duckdb(
  con,
  path,
  name = sub("\\\\..*$", "", basename(path)),
  read_only = FALSE
)
}
\arguments{
\item{con}{DuckDB connection.}

\item{path}{Path of the new DuckDB file.}

\item{name}{Name to use for the new database. Defaults the name of the file.}

\item{read_only}{Read only. Defaults to false.}
}
\value{
A single value vector with a value of zero if successful.
}
\description{
Attach an Additional DuckDB File to Connection
}
