% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_query.R
\name{collect_query}
\alias{collect_query}
\title{Collect Query from SQL File}
\usage{
collect_query(conn, path)
}
\arguments{
\item{conn}{A database connection.}

\item{path}{File path to a SQL file.}
}
\value{
A tibble containing the results of the query.
}
\description{
From a SQL file, query the database and return the results in a tibble.
}
\details{
The SQL must not end with a semicolon. Whitespace and comments are allowed.
}
