% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_duplicates.R
\name{count_duplicates}
\alias{count_duplicates}
\title{Count Duplicate Rows}
\usage{
count_duplicates(.data, ...)
}
\arguments{
\item{.data}{A \code{\link[tibble:tibble]{tibble::tibble}} or data frame. There should not be a column
with the name \code{n}. If there is, this won't work right.}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Variables to group
by.}
}
\value{
A tibble (or data frame) with columns from \code{...} and an \code{n} column
with the number of rows for each combination of \code{...}.
}
\description{
Count the number of rows and filter to show which combinations have
duplicates. This is just a wrapper around \code{\link[dplyr:count]{dplyr::count}} and
\code{\link[dplyr:filter]{dplyr::filter}}.
}
\examples{
mtcars |>
    count_duplicates(cyl)
}
