% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_duckdb.R
\name{create_table_manually}
\alias{create_table_manually}
\title{Manual DuckDB Table Creation}
\usage{
create_table_manually(con, name, value, overwrite = TRUE, temporary = FALSE)
}
\arguments{
\item{con}{DuckDB connection.}

\item{name}{Table name.}

\item{value}{The table.}

\item{overwrite}{Overwrite.}

\item{temporary}{Temporary.}
}
\description{
This is an internal function for creating new tables in DuckDB. This avoids
constructing a SQL statement each time.
}
\details{
This is only for lazy tables, such as materializing a table from DuckDB into
DuckDB.
}
\keyword{internal}
