% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographic_distance.R
\name{geographic_distance}
\alias{geographic_distance}
\title{Geographic Distance Between Two Pairs of Coordinates}
\usage{
geographic_distance(lat, long, lat_ref, long_ref, unit = c("miles", "km"))
}
\arguments{
\item{lat}{A vector of latitude coordinates.}

\item{long}{A vector of longitude coordinates.}

\item{lat_ref}{A reference latitude.}

\item{long_ref}{A reference longitude.}

\item{unit}{The output unit. Must be miles or kilometers (\code{km}). The
conversion factor is 1609.334 meters to 1 mile.}
}
\value{
A vector of the distance between to points on a globe.
}
\description{
This wraps \code{\link[geosphere:distGeo]{geosphere::distGeo}} and allows unit conversion.
}
