% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not-in.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In}
\usage{
lhs \%notin\% rhs
}
\arguments{
\item{lhs}{A vector.}

\item{rhs}{A vector.}
}
\value{
A logical vector with the same length as \code{lhs}.
}
\description{
The opposite of \code{\link{\%in\%}}.
}
\examples{
a <- c(1, 2, 3)
b <- c(1, 2)

# logical vector
a \%notin\% b

# equivalent to
!(a \%in\% b)
}
