% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-cons.R
\name{oracle_con}
\alias{oracle_con}
\title{Oracle Database Connection}
\usage{
oracle_con(user, pass, dbname)
}
\arguments{
\item{user}{Oracle username.}

\item{pass}{Oracle password.}

\item{dbname}{Database name.}
}
\value{
An ROracle database connection (\code{OraConnection}).
}
\description{
Connect to an Oracle database. Provide the username and password through
secret means.
}
\details{
Installing the \code{ROracle} package can be complicated. It requires the OCI
library. Instructions can be found on the Oracle website and require an
Oracle account. The \href{https://cran.r-project.org/package=ROracle}{version on CRAN} is outdated, see
\href{https://www.oracle.com/database/technologies/roracle-downloads.html}{Oracle}
for the most recent release.
}
