% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns_na.R
\name{select_columns_na}
\alias{select_columns_na}
\title{Select Columns That Have Missing Values}
\usage{
select_columns_na(.data, type = c("any", "all"))
}
\arguments{
\item{.data}{A tibble.}

\item{type}{A string, either "any" or "all". Using "any" will take columns
with one or more missing values. Using "all" will take only columns with
only missing values.}
}
\value{
A tibble.
}
\description{
Select Columns That Have Missing Values
}
\examples{
df <- tibble::tribble(
    ~ x, ~ y, ~ z,
    NA,   NA,   3,
    NA,    2,   3
)

select_columns_na(df, "any")

select_columns_na(df, "all")
}
