% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb-extensions.R
\name{install_duckdb_extension}
\alias{install_duckdb_extension}
\title{Install DuckDB Extensions Through HTTPS}
\usage{
install_duckdb_extension(
  duckdb_connection = duck_con(),
  extension_name,
  duckdb_version,
  platform_name
)
}
\arguments{
\item{duckdb_connection}{The DuckDB connection. Defaults the the in-memory
database.}

\item{extension_name}{Name of the extension to install. A list of extensions
can be found in the \href{https://duckdb.org/docs/stable/core_extensions/overview.html}{DuckDB documentation}.}

\item{duckdb_version}{(Only needed for installing \code{httpfs}.) The version of
DuckDB installed, don't include the 'v'. It can be found with
\code{packageVersion("duckdb")}.}

\item{platform_name}{(Only needed for installing \code{httpfs}.) See the
\href{https://duckdb.org/docs/stable/dev/building/overview.html#platforms}{platforms}
for reference. This is checked against the list of platforms from the
reference.}
}
\description{
The default core extension repository works through HTTP. There may be
network limitations on directly downloading a file in this way (such as in an
enterprise setting). To get around this, you can just add an 'S' in the URL
to make it use HTTPS. This function does that automatically for you.
}
\details{
More
\href{https://duckdb.org/docs/stable/extensions/advanced_installation_methods}{info}
about DuckDB extension installation.
}
